+++ plugins/security-metrics/src/components/shared/ViewSettingsDialog.tsx
@@ -4,78 +4,107 @@ import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogTitle from '@mui/material/DialogTitle';
 import Divider from '@mui/material/Divider';
-import FormControlLabel from '@mui/material/FormControlLabel';
+import List from '@mui/material/List';
+import ListItem from '@mui/material/ListItem';
+import ListItemText from '@mui/material/ListItemText';
 import Switch from '@mui/material/Switch';
-import Typography from '@mui/material/Typography';
-import { FilterEnum } from '../../typesFrontend';
-
-interface StarFilterProps {
-  hasStarred: boolean;
-  effectiveFilter: FilterEnum;
-  onToggleStarFilter: (
-    event: React.ChangeEvent<HTMLInputElement>,
-    checked: boolean,
-  ) => void;
-}
+import { StarFilterProps } from './ViewSettingsButton';
 
 interface Props {
   open: boolean;
   onClose: () => void;
   starFilter?: StarFilterProps;
   showTotal: boolean;
+  showOpen: boolean;
   onToggleShowTotal: (
     event: React.ChangeEvent<HTMLInputElement>,
     checked: boolean,
   ) => void;
+  onToggleShowOpen: (
+    event: React.ChangeEvent<HTMLInputElement>,
+    checked: boolean,
+  ) => void;
 }
+
+interface SettingSwitchProps {
+  title: string;
+  description: string;
+  checked: boolean;
+  disabled?: boolean;
+  onChange: (
+    event: React.ChangeEvent<HTMLInputElement>,
+    checked: boolean,
+  ) => void;
+}
+
+const SettingSwitch = ({
+  title,
+  description,
+  checked,
+  disabled = false,
+  onChange,
+}: SettingSwitchProps) => (
+  <ListItem
+    disableGutters
+    secondaryAction={
+      <Switch checked={checked} disabled={disabled} onChange={onChange} />
+    }
+  >
+    <ListItemText primary={title} secondary={description} sx={{ pr: 6 }} />
+  </ListItem>
+);
 
 export const ViewSettingsDialog = ({
   open,
   onClose,
   starFilter,
   showTotal,
+  showOpen,
   onToggleShowTotal,
+  onToggleShowOpen,
 }: Props) => {
   return (
-    <Dialog open={open} onClose={onClose} maxWidth="xs" fullWidth>
+    <Dialog open={open} onClose={onClose} maxWidth="sm" fullWidth>
       <DialogTitle>Tilpass visning</DialogTitle>
-      <DialogContent
-        sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          gap: 2,
-          pt: '12px !important',
-        }}
-      >
-        {starFilter && (
-          <>
-            <Typography variant="subtitle2" color="text.secondary">
-              Komponenter
-            </Typography>
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={starFilter.effectiveFilter === 'starred'}
-                  onChange={starFilter.onToggleStarFilter}
-                  disabled={!starFilter.hasStarred}
-                />
-              }
-              label="Vis kun stjernemarkerte komponenter"
-            />
-            <Divider />
-          </>
-        )}
 
-        <Typography variant="subtitle2" color="text.secondary">
-          Trend
-        </Typography>
-        <FormControlLabel
-          control={<Switch checked={showTotal} onChange={onToggleShowTotal} />}
-          label="Vis totalt antall sårbarheter"
-        />
+      <DialogContent>
+        <List disablePadding>
+          <SettingSwitch
+            title="Vis bare åpne sårbarheter"
+            description="Skjuler aksepterte sårbarheter fra de aggregerte visningene."
+            checked={showOpen}
+            onChange={onToggleShowOpen}
+          />
+          <Divider component="li" />
+          <SettingSwitch
+            title="Vis totalt antall i trendgrafen"
+            description="Legger til totalt antall sårbarheter i trendgrafen, i tillegg til kritiske og høye sårbarheter."
+            checked={showTotal}
+            onChange={onToggleShowTotal}
+          />
+          {starFilter && (
+            <>
+              <Divider component="li" />
+              <SettingSwitch
+                title="Vis bare stjernemarkerte komponenter"
+                description={
+                  starFilter.hasStarred
+                    ? 'Filtrerer oversikten til komponentene du har markert med stjerne.'
+                    : 'Du har ingen stjernemarkerte komponenter ennå. Marker en komponent med stjerne i katalogen for å bruke dette filteret.'
+                }
+                checked={starFilter.effectiveFilter === 'starred'}
+                disabled={!starFilter.hasStarred}
+                onChange={starFilter.onToggleStarFilter}
+              />
+            </>
+          )}
+        </List>
       </DialogContent>
-      <DialogActions sx={{ justifyContent: 'flex-end', pb: 2, pr: 2 }}>
-        <Button onClick={onClose}>Lukk</Button>
+
+      <DialogActions>
+        <Button onClick={onClose} sx={{ p: 2 }}>
+          Ferdig
+        </Button>
       </DialogActions>
     </Dialog>
   );
